+++ src/platform/fallback.rs
@@ -111,6 +111,10 @@ pub(crate) fn local_datetime() -> Option<time::PrimitiveDateTime> {
     None
 }
 
+pub(crate) fn local_datetime_at(_timestamp: u64) -> Option<time::PrimitiveDateTime> {
+    None
+}
+
 pub(crate) fn status_commands_supported() -> bool {
     false
 }